+++ src/main/java/gregicadditions/recipes/qubit/QubitProducerRecipeBuilder.java
@@ -15,7 +15,7 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 
 public class QubitProducerRecipeBuilder extends RecipeBuilder<QubitProducerRecipeBuilder> {
 
-    private long qubit;
+    private int qubit;
 
     public QubitProducerRecipeBuilder() {
     }
@@ -37,13 +37,13 @@ public class QubitProducerRecipeBuilder extends RecipeBuilder<QubitProducerRecip
     @Override
     public boolean applyProperty(String key, Object value) {
         if (key.equals("qubit")) {
-            this.qubit(((Number) value).longValue());
+            this.qubit(((Number) value).intValue());
             return true;
         }
         return true;
     }
 
-    public QubitProducerRecipeBuilder qubit(long qubit) {
+    public QubitProducerRecipeBuilder qubit(int qubit) {
         if (qubit <= 0) {
             GTLog.logger.error("qubit cannot be less than or equal to 0", new IllegalArgumentException());
             recipeStatus = EnumValidationResult.INVALID;
